sched/{core,cpufreq_schedutil}: add capacity clamping for RT/DL tasks
Currently schedutil enforce a maximum OPP when RT/DL tasks are RUNNABLE.
Such a mandatory policy can be made more tunable from userspace thus
allowing for example to define a reasonable max frequency which is
required for the execution of a specific RT/DL workload. This will
contribute to make the RT class more friendly for power/energy sensible
applications.

This patch extends the usage of capacity_{min,max} to the RT/DL classes.
Whenever a task in these classes in RUNNABLE, the capacity required is
defined by the constraints of the control group the task belong to.

Change-Id: I5e4b20b008c4d100677c4fa411ad8b209fd262d1
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>