[WIP] arm,arm64,drivers: Add infrastructure for setting SD_ASYM_CPUCAPACITY
This adds code for detecting asymmetrical-capacity systems and setting the
SD_ASYM_CPUCAPACITY flag at the appropriate sched_domain level.

On some systems (such as those where each CPU in the system is architecturally
identical but differ in their range of available frequencies), the fact of
asymmetry cannot be determined at initial boot. This patch adds code to handle
these cases by updating the flag when cpufreq has come up and when userspace
updates the cpu_capacity sysfs files.

WIP: The update_sched_flags implementation is undesirable as it requires
defining an arch_ function in drivers/. This is necessary because it uses a
global variable and a the arch_update_cpu_topology callback to communicate with
rebuild_sched_domains (which is also undesirable in itself). A later version
will likely rely on adding a "force" flag to rebuild_sched_domains.

Signed-off-by: Brendan Jackman <brendan.jackman@arm.com>