sched/fair.c: Task contribution used to find lowest utilization
This should fix a potential over-packing issue.
In some cases, the new_util of a task can be the same on several
CPUs. This causes an issue because the target_util is only updated
if the current new_util is strictly smaller than target_util.

To fix that, the cpu_util_wake() return value is used alongside the
new_util value. If two CPUs compute the same new_util value,
we'll now also look at their cpu_util_wake() return value. In this
case, the CPU that last ran the task will be chosen in priority.

Change-Id: I5497ba791f31fdce26b9a252eac2c00fb0aa3a4e
Signed-off-by: Valentin Schneider <valentin.schneider@arm.com>