sched/fair: avoid subtracting unrelated signals when using WALT
we use cpu_util_wake to subtract the tasks blocked load from the
cpu it previously ran on, however this operation only makes sense
for PELT where task and cpu utilization are both decayed at the
same rate. When WALT is in use, cpu utilization is the utilization
in the previous completed window, whilst the task utilization
ignores fully idle windows. There seems to be no way to have a
decently accurate estimate of how much (if any) utilization from
this task remains on the prev cpu.

Instead, just return cpu_util when we're using WALT.

Change-Id: Idb153f28fc44e6dddf7f5ca6ff6e9fdf843c482a
Reported-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>