PM / devfreq: arm: add devfreq driver for ARM DSU cache portion control
The cache in the ARM DynamIQ Shared Unit (DSU) supports partial power down
by splitting the cache into an implementation-specific number of portions.
The DSU portion control DevFreq driver provides an energy-cost justified
demand-driven policy for controlling the number of portions enabled.

The driver also provides a specific DevFreq governor that will implement
the desired energy-cost-justification policy. This policy maps the number
of portions enabled to frequency (1 portion == 1Hz).
The driver does not control the clock frequency of the cache.
This version of the code is not suitable for use with the simple
ondemand governor.

Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>