EXPERIMENTAL: sched/fair: Reduce balance interval to 0 if we have a misfit task
It is quite a common occurrence that we have a misfit task (a task
running on a CPU which is close to the capacity limit of that CPU)
in the system AND we have recently done a load balance operation.
Normal load-balance rate limiting means we can leave those tasks on
the restrictive CPU for a few hundred ms before we move them.

This patch reduces the balance interval to zero when there are such
tasks running.

Change-Id: I31e8d1225fad043b099c1b4057998cbf9d2a2bed
Signed-off-by: Chris Redpath <chris.redpath@arm.com>