sched/cpufreq_schedutil: use util_est for OPP selection
When the schedutil looks at the CPU utlization, the current PELT value for
that CPU is returned straightway. In certain scenarios this can have
undesired side effects on frequency selection.

For example, since the task utilization is decayed at wakeup time, when a
long sleeping big task is enqueued it does not add immediately a
significant contribution to the target CPU. This introduces some latency
before schedutil will be able to detect the best frequency required by that
task. Moreover, since the PELT signal building up timing is function of the
current frequency, starting from a lower one will increase even more the
time required to reach the actual frequency to better serve the task
requirements.

In order to reduce this kind of latencies, this patch integrate the usage
of the CPU's estimated utlization in the sugov_get_util function.

The estimated utilization of a CPU is defined to be the maximum between its
PELT's utilization and the sum of the estimated utilization of the tasks
currently RUNNABLE on that CPU.
This allows to properly represent the expected utilization of a CPU which,
for example it has just got a big task running since a long sleep period,
thus allowing to select since its activation the best frequency to run that
task.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Paul Turner <pjt@google.com>
Cc: Vincent Guittot <vincent.guittot@linaro.org>
Cc: Morten Rasmussen <morten.rasmussen@arm.com>
Cc: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Cc: linux-kernel@vger.kernel.org
Cc: linux-pm@vger.kernel.org