sched/rt: add utilization tracking
schedutil governor relies on scheduler's util_avg to choose the OPP when cfs
tasks are running. When the CPU is overloaded by cfs and rt tasks, cfs tasks
are preempted by rt tasks and in this case util_avg reflects the remaining
capacity that is used by cfs task but not what cfs want to use. In such case,
schedutil can select a lower OPP whereas the CPU is overloaded.In order to have
a more accurate view of the utilization of the CPU, we track the utilization
that is "stolen" by RT tasks.