arm64: dts: Remove "arm,vexpress" comatible string from FVP dts files
FVP Base models aren't really that compatible with Versatile Express and
the only thing in Linux that tests for that string is the platform code
which registers smp ops in 32-bit kernels. As we use PSCI for CPU
bringup remove this compatible string and let 32-bit kernels work wore
reliably.

Signed-off-by: Jon Medhurst <tixy@linaro.org>