cpufreq/sched: Consider max cpu capacity when choosing frequencies
When using schedfreq on cpus with max capacity significantly smaller than
1024, we were selecting an incorrect OPP as we were scaling the frequency
as if the max capacity achievable was 1024 rather than the max for that
particular cpu or group. This could result in a cpu being stuck at the
lowest OPP and unable to generate enough utilisation to climb out if the
max capacity is significantly smaller than 1024.

Instead, unscale the capacity to be in the range 0-1024 so that when we
later select a frequency, we get the correct one.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>