sched: cpufreq: Allow one cpu to update the util of another CPU
Load tracking updates may happen for CPU X on another CPU Y. To allow
for faster updates, have CPU X update CPU Y's utilization, so that
the governor may respond faster.

This is necessary for WALT migration updates, otherwise a migration
invoked load update at the beginning of a window may only be known
to the governor in the next window. Considering a window size of 20ms
this is not a good idea; the task deserves the performance level
it has already earned on its previous CPU/cluster.

Change-Id: I2439ed41a0d14b92ddaf2c7ba7b2ff9c35af7370
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>