arm64: dts: hi3660: Add EAS energy model provide big.LITTLE info
This patch is to enable energy model parameters for P-states and C-states. So
use it to enable energy aware scheduling.

Each OPP has been benchmarked with dhry2.1 while the board was cooled with a
heatsink and a fan to prevent any thermal mitigation. The DDR frequency was also
locked at 400MHz. The results show that the A53 capacity values can be slightly
increased.

Idle costs are trickier to analyse on the HiKey960, and Leo rightfully pointed
out that they should not be greater than active costs - this is actually the
case with the current values, as the A73 cluster idle costs are greater than the
lowest cluster active costs.

As such, each idle value is defined as: min(measured_value, min(active_costs))

Signed-off-by: Leo Yan <leo.yan@linaro.org>
Signed-off-by: John Stultz <john.stultz@linaro.org>
Signed-off-by: Valentin Schneider <valentin.schneider@arm.com>