drivers base/arch_topology: rebuild sd hierarchy during smp/hmp switch
The setup of the scheduler domain (sd) hierarchy differs between an smp
(symmetric multi-processing) and an hmp (heterogeneous multi-processing)
system.
For the latter one the sd topology flag SD_ASYM_CPUCAPACITY has to be
set on the sd's in which the cpu capacity of the spaned cpus differs.

The problem is that an hmp system can not be detected until the
capacity-dmips-mhz dt-bindings have been read. At this moment, the
initial (default smp) sd has already been built. So we need to use
rebuild_sched_domains() to switch towards hmp during startup or between
smp and hmp in case that the cpu capacity is changed via the
sysfs interface /sys/devices/system/cpu/cpuX/cpu_capacity.

By default, i.e. during boot process, we assume we run an smp system
(topology_hmp equal 0).

Then we check:

(1) if we have to switch to hmp from the initial smp system for:

(a) an hmp system purely based on capacity-dmips-mhz dt-bindings in:

    parse_dt_topology()
    -> topology_normalize_cpu_scale()
       -> check_rebuild_sched_domains()

(b) an hmp system based on capacity-dmips-mhz dt-bindings and cpufreq
    rescaling based on max supported frequency in:

    init_cpu_capacity_callback()
    -> topology_normalize_cpu_scale()
       -> check_rebuild_sched_domains()

(2) if we have to switch to hmp or smp for:

(a) cpu_rq(cpu)->cpu_capacity is changed via the sysfs interface:

    cpu_capacity_store()
    -> check_rebuild_sched_domains()

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>