drivers base/arch_topology: rebuild sd hierarchy during smp/hmp switch
The setup of the scheduler domain (sd) hierarchy between an smp and an
hmp system differs.
For the latter one the sd topology flag SD_ASYM_CPUCAPACITY has to be
set on those sd's in which the cpu capacity of the spaned cpus differs.

The problem is that an hmp system can not be detected until the
capacity-dmips-mhz dt-bindings have been read. At this moment, the
initial (default smp) sd has already been built. So we need to use
rebuild_sched_domains() to switch towards hmp during startup or between
smp and hmp in case that the cpu capacity is changed via the
sysfs interface /sys/devices/system/cpu/cpuX/cpu_capacity.

By default, i.e. during boot process, we assume we run an smp system
(topology_hmp equal 0).

(1) Check if we have to switch to hmp from the initial smp system for:

(a) hmp system purely based on capacity-dmips-mhz dt-bindings in:

    parse_dt_topology()
    -> topology_normalize_cpu_scale()
       -> check_rebuild_sd_needed()

(b) hmp system based on capacity-dmips-mhz dt-bindings and cpufreq
    rescaling based on max supported frequency in:

    init_cpu_capacity_callback()
    -> topology_normalize_cpu_scale()
       -> check_rebuild_sd_needed()

(2) Check if we have to switch to hmp or smp for:

(a) cpu_rq(cpu)->cpu_capacity is changed via the sysfs interface:

    cpu_capacity_store()
    -> check_rebuild_sd_needed()

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>