firmware: arm_scmi: add initial support for sensor protocol
The sensor protocol provides functions to manage platform sensors, and
provides the commands to describe the protocol version and the various
attribute flags. It also provides commands to discover various sensors
implemented and managed by the platform, read any sensor synchronously
or asynchronously as allowed by the platform, program sensor attributes
and/or configurations, if applicable.

This patch adds support for most of the above features.

Change-Id: I9ad5c79414f71de29fcc3f2b5515c50f01753ef3
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>