sched/{fair,tune}: trace energy_diff payoff reason
The energy_diff function is used to ultimately accept or reject a
certain scheduling candidate in alternative to run a task in its
previous CPU. This accept/reject result of a scheduling candidate can be
due to different reasons, some of them involving the SchedTune filtering
and thus a specific trade-off on performance vs energy

To better support trace analysis of energy_diff decisions, it can be
useful to report what was the ultimate heuristic which has been used to
accept or reject a scheduling candidate.

This patch extends the sched_energy_diff() trace event to report the
heuristic which has been used to decide whether to accept or reject a
scheduling candidate.

Change-Id: I0ebb7c32364e3cafa5816e9f8b79eed29292aad8
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>