drm/edid: introducing EDID retry
Due to some hardware issues on Juno platform, the received EDID may be broken.
The only solution we have so far is to retry to get it until it works or until
we reach the limit set by CONFIG_DRM_EDID_RETRY or module parameter
drm.edid_retry.

Jira: MPDBSP-1369
Change-Id: I59a9500435def4ef6e5b18634a4b2d30f7dda222
Signed-off-by: Vincent Donnefort <vincent.donnefort@arm.com>