dt-bindings: mailbox: add support for mailbox client shared memory
Many users of the mailbox controllers depend on the shared memory between the
two end points to exchange the main data while using simple doorbell mechanism
to alert the end points of the presence of a message.

This patch defines device tree bindings to represent such shared memory in a
generic way.

Change-Id: I798369f83f91ad1be7e50469d0722c84caecc97a
Acked-by: Rob Herring <robh+dt@kernel.org>
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>