dt-bindings: mailbox: add support for mailbox client shared memory
Many users of the mailbox controllers depend on the shared memory between the
two end points to exchange the main data while using simple doorbell mechanism
to alert the end points of the presence of a message.

This patch defines device tree bindings to represent such shared memory in a
generic way.

Acked-by: Rob Herring <robh+dt@kernel.org>
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>