arm,arm64,drivers,sched: allow to set SD_ASYM_CPUCAPACITY on all sd levels
To be able to set SD_ASYM_CPUCAPACITY on every sd level and not only on
DIE create a topology flag function with a cpumask parameter:

  int topology_flags(const struct cpumask *mask)

For backward compatibility, the existing DIE level topology flag
function topology_cpu_flags(void) calls this function with the DIE level
cpumask.

Moreover, this function can now be called to detect heterogeneity inside
a core (SMT) or inside a cluster (future ARM DynamIQ).

The task scheduler does not call the sched domain (sd) level flag
function with the cpu id as an argument so we have to derive the cpumask
from the current cpu. This is still possible since the scheduler
requires symmetric (i.e. if an sd flag is set for one cpu it has to be
set for any cpu) sd hierarchy setups.

This patch enforces the rule that the SD_ASYM_CPUCAPACITY flag is set
on the highest sd in which the cpu capacity values differ in sd_init().

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>