arm,arm64,drivers: make topology_flags() usable for other sched domain flags and levels
To be able to set other topology sched domain (sd) flags than
SD_ASYM_CPUCAPACITY on every sd level and not only on DIE add a cpumask
parameter to topology_flags().

This requires also to change to implementation to detect if the
SD_ASYM_CPUCAPACITY sd flag has to be set on DIE level.

The support of existing big.Little systems is still provided. Moreover
the new implementation can also detect heterogeneity inside a cluster
(i.e. on MC sd level) and so provide support for future ARM DynamIQ
cpus.

Since topology_flags() now has a cpumask argument its use in
arm[,64]_topology[] has to be slightly adapted.

The task scheduler does not call the sched domain (sd) level flag
function with the cpu id as an argument so we have to derive the cpumask
from the current cpu. This is still possible since the scheduler
requires symmetric (i.e. if an sd flag is set for one cpu it has to be
set for any cpu) sd hierarchy setups.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>