arm,arm64,drivers : set SD_SHARE_CAP_STATES sched domain flag
The EAS scheduler needs to know about clock domains, i.e. which cpus are
sharing a clock source.

This patch extends the functionality in check_rebuild_sched_domains() to
force a rebuild of the scheduler domains (sd) the first time cpufreq is
available. This is necessary to be able to retrieve the cpufreq 'related
cpus' mask via cpufreq_cpu_get() to compare it with the mask of cpus
spanning an sd.

The generic topology flag function topology_flags() is extended to set
the SD_SHARE_CAP_STATES flag in case the cpufreq 'related cpus' mask is
equal to the mask of cpus spanning this sd.

Since cpus of a MC sd level can share the clock domain exclusively, add
a specific topology flag function topology_core_flags() for the MC sd
level.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>