arm: topology: remove cpu_efficiency
Remove the old cpu_efficiency/dt property 'clock-frequency' based
solution to set cpu capacity which was only working for Cortex-A15/A7
arm big.LITTLE systems.

This means that the new 'capacity-dmips-mhz' based solution is now used
as the default one which is shared between arm and arm64 and works for
all kind of big.LITTLE systems.

Cc: Russell King <linux@arm.linux.org.uk>
Cc: Vincent Guittot <vincent.guittot@linaro.org>
Cc: Juri Lelli <juri.lelli@arm.com>
Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>