cpufreq: remove max freq invariance from scale_freq_capacity()
With this patch, scale_freq_capacity() is now only the setter of the
freq invariance.
That's why it also deletes the call to scale_freq_capacity() in
cpufreq_set_policy().
Change the parameter list of scale_freq_capacity() to (1) set of cpus
involved (cpumask), (2) current frequency, (3) maximum possible
operating frequency.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>