kernel/sched: add relative energy normalization
SchedTune normalize energy differences with respect to the maximum
possible energy consumption. This allows to get an index for energy
variations which can be easily represented in the PE filtering space.
However, in many cases EAS is trying to move small amount of energy from
one CPU to another.

Using this approach. some possible energy optimizations appears as
negligible energy differences in absolute terms. Moreover, while
energy variations are evaluated in absolute terms, the performance index
variations for the same tasks are evaluated using a more local/relative
metric which considers just at the single task.

To solve these two possible issues, an alternative definition of the
energy variation index is possible where we always normalize the energy
diff with respect to the maximum of the energy consumed in the two
possible candidate solutions.

This patch introduces a relative energy normalization with an additional
sched feature flag which allows to choose between relative and absolute
normalization. Such a flag it's useful to better test/compare the two
different normalization policies.

Change-Id: I6c988298527fdeb19a1eb5d907c8325080694bf1
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
(minor cherry-pick conflicts)
Signed-off-by: Chris Redpath <chris.redpath@arm.com>