msm: mdss: allow transitional alpm commands
Current ALPM commands assume that it's entering ALPM mode from ON state.
This may not always be true when transitioning between different
brightness levels. In such cases there are some commands that could be
avoided in order to reduce flicker or other artifacts since we can
assume the display is already in ALPM mode. Allow for a different set of
commands that can be used during transitions.

Bug: 63983815
Bug: 63997216
Change-Id: I2531f2b9ebacf682464f51f7fc9cfa2d53edd718
Signed-off-by: Adrian Salido <salidoa@google.com>