sched/fair: add low_util support for prefer_idle
Currently, for prefer_idle tasks, we pick the first IDLE CPU found in a
big-to-LITTLE exploration of the oneline CPUs. This approach tends to
favour big CPUs over LITTLE since for boosted tasks we explore CPUs
starting from higher capacity ones.
However, in lightly loaded scenarios, it can still make sense to select
a more energy-efficient IDLE CPU to run latency sensitive tasks.

This patch introduces the concept of "low utilization mode" which
corresponds to all online CPUs with a utilization smaller then a
predefined threshold high_cpu_util (e.g. 1/4 of the maximum capacity).

With this new concept, once we explore CPUs for a prefer_idle task,
while expored CPUs are verified to be in low_util_mode, we do not return
with the first IDLE CPU found. Instead we keep exploting the remaining
CPUs to update a low_util_cpu candidate. This is the IDLE CPU with the
lower capacity_orig in the system, thus likely an IDLE and
energy-efficient CPU.

During CPUs exploration, if a CPU is found to be above the high_cpu_util
utilization value, we immediately return the first identified IDLE cpu.
Otherwise, if all CPUs are below that threshold, the low_util_cpu is the
returned target cpu.

In order to reduce wakeup latencies, this new version still skip
updating best_active_cpus as soon as an IDLE CPU has been found.

Change-Id: I25adc0f85719457529969c13b96dac03acb57431
Suggested-by: Joel Fernandes <joelaf@google.com>
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: Quentin Perret <quentin.perret@arm.com>