misc: mnh: fix handling of pcie suspend in failure cases
If easel goes down while PCIe is in L1 state, the AP PHY might not
receive the linkdown irq. Modify the PCIe suspend sequence to suspend
the link without reading in the configuration space. This puts the MSM
PCIe driver to be in the correct state when we call resume next.

Bug: 64679990
Change-Id: Ie91a92ba1afcb93126b3f66cf0e24ebfc6d5720f
Signed-off-by: Trevor Bunker <trevorbunker@google.com>