PM / devfreq: dsu-pctrl: consider a portion as 50% of the cache
The data ram can only be powered on/off in portions of 50% of the cache
while the tag rams can be powered on/off in portions of 25% of the cache.
This results in the vast majority of the leakage being no different at 1
and 2 portions, and likewise at 3 and 4 portions. If we select 1 or 3
portions we're losing a quarter of the cache for almost no leakage
reduction.

The current powerdown register interface is linked to the tag ram split
of portions, but for performance reasons we consider one portion as the
equivalent of 50% of the cache and write the powerdown register
accordingly.

Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>